+++ changelog.md
@@ -1,9 +1,16 @@
-# Changelog for PSCalendar
+# Change Log for PSCalendar
+
+## v1.2.0
+
++ Added validation for `-Month` (Issue #1)
++ Made `-Year` a positional parameter in the 2nd position
++ some minor code cleanup
++ Added a private helper function to enumerate month names (Issue #2)
 
 ## v1.1.0
 
 + Fixed dot sourcing of calendar functions script.
-+ Updated `README.me`
++ Updated `README.md`
 + Verified functionality on PowerShell Core
 + Published to the PowerShell Gallery